Adds clickable graph values as links.
Adds a list of links below the main video, allowing users to
click on specific data points from the graph.  These links open
the corresponding YouTube video at the relevant timestamp. Skips
the latest entry as its rendered in the iframe, and hides the
list if empty.

Also, enables dotenv to pick up environment variables, and logs
the data source URL for debugging.